Add support for application protocol detection
We need to support tracing for common protocols like HTTP/1, HTTP/2, gRPC, and support pass-through for the rest of the protocols until we add support for them.

This PR implements a custom Envoy listener filter that reads the first
few kilobytes of every request, detects the protocol and annotates the
request with the detected protocol for Envoy filter chains to use the
information later.

Issue: A0-167